+++ src/NeatMapper.EntityFrameworkCore/Mappers/EntityFrameworkCoreBaseMapper.cs
@@ -17,7 +17,6 @@ using System.Collections;
 using System.Linq.Expressions;
 using System.Collections.Concurrent;
 using System.Threading;
-using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore.ChangeTracking;
 namespace NeatMapper.EntityFrameworkCore {
 	/// <summary>